#63 Extract message handling into handlers
